WIP Move details to tabs
The details of the calculations make for a lot of info in one component. It should be easier to show everything useful to a particular category if it has its own tab.
